p3701: update pwm-fan cooling levels of AGX Orin
The cooling levels used to be ignored as the userspace nvfancontrol
service took control on the fan speed based on the given profiles and
the thermal groups, etc. So, the pwm-fan cooling levels were not taken
care of in the past.

But nowadays it's observed that some boards get extremely hot during
flashing and sometimes causes the flash failure. To dissipate
the heat during l4t initrd flash, the pwm-fan module is loaded into the
initrd image. To strike a balance between the cooling effect and the
noise caused by the fan, this change updates the pwm-fan cooling levels
to make the fan spin at 50% speed by default.

Bug 3872696

Change-Id: I242e11758644c24f3a3498b3b3951e1b86478582
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/platform/t23x/concord-dts/+/2857142
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Yi-Wei Wang <yiweiw@nvidia.com>
Reviewed-by: Rajkumar Kasirajan <rkasirajan@nvidia.com>
Tested-by: Yi-Wei Wang <yiweiw@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>